add modal sucesse and error in project

+++ app/Http/Controllers/ServiceController.php
@@ -2,12 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Constantes\Constantes;
 use App\Models\Service;
-use App\Models\ServiceOrder;
 use App\Validator\ServiceValidator;
 use App\Validator\ValidationException;
 use Illuminate\Http\Request;
-
+use Exception;
 class ServiceController extends Controller
 {
     public function createView(Request $request){
@@ -17,14 +17,26 @@ class ServiceController extends Controller
     public function create(Request $request){
         try {
             ServiceValidator::validate($request->all());
-            $service = new Service($request->all());
-            $service->save();
-            return redirect('/service');
+            try {
+                $service = new Service($request->all());
+                $service->save();
+
+                return redirect('/service')
+                    ->with('mensagem', Constantes::SUCESSO_CREATE_SERVICE);
+
+            } catch (Exception $ex){
+                return redirect('/service/create')
+                    ->with('mensagem', Constantes::ERROR_CREATE_SERVICE);
+            }
+
         } catch (ValidationException $ve){
             return redirect('/service/create')
+                ->with('mensagem', Constantes::ERROR_CREATE_SERVICE)
                 ->withErrors($ve->getValidator())
                 ->withInput();
         }
+
+
     }
 
     public function updateView(Request $request){
@@ -35,13 +47,21 @@ class ServiceController extends Controller
     public function update(Request $request){
         try {
             ServiceValidator::validate($request->all());
-            $service = Service::find($request->idService);
-            $service->descricao = $request->descricao;
-            $service->valor = $request->valor;
-            $service->update();
-            return redirect('/service');
+            try {
+                $service = Service::find($request->idService);
+                $service->descricao = $request->descricao;
+                $service->valor = $request->valor;
+                $service->update();
+                return redirect('/service')
+                    ->with('mensagem', Constantes::SUCESSO_UPDATE_SERVICE);
+            } catch (Exception $exception){
+                return redirect('/service/update/'.$request->idService)
+                    ->with('mensagem', Constantes::ERROR_UPDATE_SERVICE);
+            }
+
         } catch (ValidationException $ve){
             return redirect('/service/update/'.$request->idService)
+                ->with('mensagem', Constantes::ERROR_UPDATE_SERVICE)
                 ->withErrors($ve->getValidator())
                 ->withInput();
         }
@@ -49,9 +69,19 @@ class ServiceController extends Controller
     }
 
     public function delete(Request $request){
-        $service = Service::find($request->idService);
-        $service->delete();
-        return redirect('/service');
+        try{
+            $service = Service::find($request->idService);
+            if ($service == null){
+                throw new Exception(Constantes::ERROR_DELETE_SERVICE);
+            }
+            $service->delete();
+            return redirect('/service')
+                ->with('mensagem', Constantes::SUCESSO_DELETE_SERVICE);;
+        } catch (Exception $ex){
+            return redirect('/service')
+                ->with('mensagem', Constantes::ERROR_DELETE_SERVICE);
+        }
+
     }
 
     public function list(Request $request){
